refactor: 불필요한 코드 삭제, 네이밍 수정
