+++ OpenMarket/OpenMarket/Extension/Data+Extension.swift
@@ -2,7 +2,7 @@
 //  Data+Extension.swift
 //  OpenMarket
 //
-//  Created by 조민호 on 2022/05/11.
+//  Created by  Red, Mino on 2022/05/11.
 //
 
 import Foundation
@@ -13,7 +13,7 @@ extension Data {
             let decoded = try JSONDecoder().decode(T.self, from: self)
             return .success(decoded)
         } catch {
-            return .failure(NetworkError.emptyData)
+            return .failure(NetworkError.decodeError)
         }
     }
 }
