+++ OpenMarket/OpenMarket/Model/DTO/ProductsListDTO.swift
@@ -5,8 +5,6 @@
 //  Created by Red, Mino on 2022/05/10.
 //
 
-import Foundation
-
 struct ProductsListDTO: Encodable {
     let pageNumber: Int
     let perPages: Int
