+++ OpenMarket/OpenMarketModelTests/NetworkProductListTest.swift
@@ -8,7 +8,7 @@
 import XCTest
 @testable import OpenMarket
 
-class NetworkProductsModelTest: XCTestCase {
+class NetworkProductListTest: XCTestCase {
     var sut: APIProvider<Products>!
     
     override func setUpWithError() throws {
@@ -23,7 +23,7 @@ class NetworkProductsModelTest: XCTestCase {
     func test_APIProvider_getProductsList_request를호출할때_예상값을반환() {
         // given
         let promise = expectation(description: "")
-        let endpoint = EndPoints.getProductsList(pageNumber: 1, perPages: 10)
+        let endpoint = EndPointStorage.getProductsList(pageNumber: 1, perPages: 10)
         
         // when
         sut.request(with: endpoint) { result in
